refactor: split imap/client.py (561 lines) into package
Extracted SyncResult and MessageStore into _sync_types.py.
Extracted email parsing into _message_parser.py.
Extracted message ops into _message_ops.py.
client.py now only contains IMAPClient class (308 lines).

All imports updated for backward compatibility.
Issue: #11
